exportovane klase za ostale dokumente
